Feat/attack goal (#77)
* Add KnightEntity and add to world

* Removed unnecessary includes, changed internal variablenames

* Accidently renamed instance to _instance in comments

thought it should not be lowercase, but seem we have different standards? so changed it back but comments were not changed back

* Only select entities from player 1

* Added player to globals and clean-up

player_id is now a const int in globals.cpp. Deleted forward declerations when inheriting

* Missed 1 hardcoded player id, moved logic, changed < to "

* updated followpathgoal, changed path to a stack instead of vector

* added more goals and some other improvements

* small fixes

* improved following/engaging

* Add KnightEntity and add to world

* Removed unnecessary includes, changed internal variablenames

* Accidently renamed instance to _instance in comments

thought it should not be lowercase, but seem we have different standards? so changed it back but comments were not changed back

* Only select entities from player 1

* Added player to globals and clean-up

player_id is now a const int in globals.cpp. Deleted forward declerations when inheriting

* Missed 1 hardcoded player id, moved logic, changed < to "

* updated followpathgoal, changed path to a stack instead of vector

* added more goals and some other improvements

* small fixes

* improved following/engaging

* removed unused stuff

* include fixes