Feat/building costs resources & Enemy-player (#78)
* Add KnightEntity and add to world

* Removed unnecessary includes, changed internal variablenames

* Accidently renamed instance to _instance in comments

thought it should not be lowercase, but seem we have different standards? so changed it back but comments were not changed back

* Only select entities from player 1

* Added player to globals and clean-up

player_id is now a const int in globals.cpp. Deleted forward declerations when inheriting

* Missed 1 hardcoded player id, moved logic, changed < to "

* Deleted unnecessary if-statement

* not working yet

* Cleaned and refactored resourcetypes, castle costs 5 wood

* Commenting resourcetypes RESTAURANT and CAMPFIRE caused problems, uncommmented...

* < to "

* Applied snake casing in player.cpp

* Changed < to " in includes

* Deleted class variables that aren't used

* Bugfix, able to add building on top of building

Implementiof of SubtractCosts.... state was wrong, it's now part of placingbuildings, also changed the set_transparent_or_border with 2 required parameters, so it's more clear how to use the method